+++ R/RNAdistribution.R
@@ -128,7 +128,8 @@
 #' @importFrom ggplot2 "labs"
 #'
 RNAdistribution  <- function (data, samples = NULL, style, 
-                              facet = TRUE, facet.arrange = 3, colour = "black", together = TRUE, 
+                              facet = TRUE, facet.arrange = 3, colour = "black", 
+                              together = TRUE, 
                               consensus = FALSE, relative = FALSE) {
   if (base::missing(data) || !base::inherits(data, c("data.frame"))) {
     stop("data must be a data frame, see ?help for more details")
@@ -139,13 +140,15 @@ RNAdistribution  <- function (data, samples = NULL, style,
     if (!relative == FALSE) {
       x <- x %>% dplyr::mutate(freq = n/sum(n))
       p1 <- ggplot2::ggplot(x, ggplot2::aes(x = DicerConsensus, 
-                                            y = freq, group = 1)) + ggplot2::geom_point() + 
+                                            y = freq, group = 1)) + 
+        ggplot2::geom_point() + 
         ggplot2::geom_line() + ggplot2::theme_classic() + 
         ggplot2::xlab("RNA Class") + ggplot2::ylab("Relative frequency")
     }
     else 
       p1 <- ggplot2::ggplot(x, ggplot2::aes(x = DicerConsensus, 
-                                               y = n, group = 1)) + ggplot2::geom_point() + ggplot2::geom_line() + 
+                                               y = n, group = 1)) + 
+        ggplot2::geom_point() + ggplot2::geom_line() + 
       ggplot2::theme_classic() + ggplot2::xlab("RNA Class") + 
       ggplot2::ylab("Counts")
     out <- list(plot = p1, data = x)
@@ -165,16 +168,13 @@ RNAdistribution  <- function (data, samples = NULL, style,
     # beware that any tables which do not have the required columns
     if (base::inherits(counts.df, c("list"))) {
       required_columns <- c("20", "21", "22", "23", "24", "N")
-      
       for (i in seq_along(counts.df)) {
         table_i <- counts.df[[i]]  # current table
-        
         if (length(names(table_i)) < 6) {
           # columns missing from the table
           missing_columns <- setdiff(required_columns, names(table_i))
           # add missing columns to table, and assign a value of 0
           table_i[missing_columns] <- 0
-          
           # check order:
           if (!identical(names(table_i), required_columns)) {
             # match order to required order
@@ -186,7 +186,8 @@ RNAdistribution  <- function (data, samples = NULL, style,
       }
       # transpose, make dataframe of results. 
       counts.df <- data.frame(t(do.call(rbind, counts.df)))
-    } else if (!base::inherits(counts.df, c("list"))) {
+    } else 
+      if (!base::inherits(counts.df, c("list"))) {
       counts.df <- data.frame(counts.df)
     }
     
@@ -199,30 +200,29 @@ RNAdistribution  <- function (data, samples = NULL, style,
     if (style == "bar") {
       plist = sapply(names(counts.df)[-grep("Class", names(counts.df))], 
                      function(col) {
-                       ggplot2::ggplot(counts.df, ggplot2::aes_string(x = "Class", 
-                                                                      y = col)) + ggplot2::geom_bar(stat = "identity", 
-                                                                                                    fill = colour) + ggplot2::theme_classic() + 
-                         ggplot2::labs(title = col, x = "RNA Class", 
-                                       y = "Count")
-                     }, simplify = FALSE)
+                       ggplot2::ggplot(counts.df, 
+                                       ggplot2::aes_string(x = "Class",y =col))+ 
+                         ggplot2::geom_bar(stat = "identity",fill = colour) + 
+                         ggplot2::theme_classic() + 
+                         ggplot2::labs(title = col, x = "RNA Class",y = "Count")
+                     }, 
+                     simplify = FALSE)
       sn <- names(plist)
       if (facet == TRUE) {
         if (is.null(samples)) {
-          p <- ggplot2::ggplot(tidyr::gather(counts.df, 
-                                             key, Count, -Class), ggplot2::aes(Class, 
-                                                                               Count)) + ggplot2::geom_bar(stat = "identity", 
-                                                                                                           fill = colour) + ggplot2::theme_classic() + 
-            ggplot2::facet_wrap(~key, scales = "free_y", 
-                                ncol = facet.arrange)
+          p <- ggplot2::ggplot(tidyr::gather(counts.df, key, Count, -Class), 
+                               ggplot2::aes(Class, Count)) + 
+            ggplot2::geom_bar(stat = "identity", fill = colour) + 
+            ggplot2::theme_classic() + 
+            ggplot2::facet_wrap(~key, scales = "free_y", ncol = facet.arrange)
         }
         else if (!is.null(samples)) {
-          counts.df <- counts.df %>% select(!all_of(samples))
-          p <- ggplot2::ggplot(tidyr::gather(counts.df, 
-                                             key, Count, -Class), ggplot2::aes(Class, 
-                                                                               Count)) + ggplot2::geom_bar(stat = "identity", 
-                                                                                                           fill = colour) + ggplot2::theme_classic() + 
-            ggplot2::facet_wrap(~key, scales = "free_y", 
-                                ncol = facet.arrange)
+          counts.df <- counts.df %>% select(all_of(samples))
+          p <- ggplot2::ggplot(tidyr::gather(counts.df, key, Count, -Class),
+                               ggplot2::aes(Class, Count)) + 
+            ggplot2::geom_bar(stat = "identity", fill = colour) + 
+            ggplot2::theme_classic() + 
+            ggplot2::facet_wrap(~key, scales = "free_y", ncol = facet.arrange)
         }
         
         out <- list(plot = p, data = counts.df)
@@ -243,9 +243,11 @@ RNAdistribution  <- function (data, samples = NULL, style,
       if (together == TRUE) {
         if (is.null(samples)) {
           counts.df_melt <- data.table::melt(counts.df, id.vars = "Class")
-          p <- ggplot2::ggplot(counts.df_melt, ggplot2::aes(Class, 
-                                                       value, col = variable, group = 1)) + ggplot2::geom_point() + 
-            ggplot2::geom_line() + ggplot2::theme_classic() + 
+          p <- ggplot2::ggplot(counts.df_melt, ggplot2::aes(Class, value, 
+                                                            col = variable, 
+                                                            group = 1)) + 
+            ggplot2::geom_point(colour = colour) + 
+            ggplot2::geom_line(colour = colour) + ggplot2::theme_classic() + 
             ggplot2::xlab("RNA Class") + ggplot2::ylab("Counts") + 
             ggplot2::labs(color = "Samples")
           out <- list(plot = p, data = counts.df)
@@ -255,8 +257,10 @@ RNAdistribution  <- function (data, samples = NULL, style,
           counts.df <- counts.df %>% select(!all_of(samples))
           counts.df <- data.table::melt(counts.df, id.vars = "Class")
           p <- ggplot2::ggplot(counts.df, ggplot2::aes(Class, 
-                                                       value, col = variable, group = 1)) + ggplot2::geom_point() + 
-            ggplot2::geom_line() + ggplot2::theme_classic() + 
+                                                       value, col = variable, 
+                                                       group = 1)) + 
+            ggplot2::geom_point(colour = colour) + 
+            ggplot2::geom_line(colour = colour) + ggplot2::theme_classic() + 
             ggplot2::xlab("RNA Class") + ggplot2::ylab("Counts") + 
             ggplot2::labs(color = "Samples")
           out <- list(plot = p, data = counts.df)
@@ -266,35 +270,44 @@ RNAdistribution  <- function (data, samples = NULL, style,
         if (together == FALSE) {
         if (facet == TRUE) {
           if (is.null(samples)) {
-            p <- ggplot2::ggplot(tidyr::gather(counts.df, 
-                                               key, Count, -Class), ggplot2::aes(Class, 
-                                                                                 Count, group = 1)) + ggplot2::geom_point() + 
-              ggplot2::geom_line() + ggplot2::theme_classic() + 
+            p <- ggplot2::ggplot(tidyr::gather(counts.df,key, Count, -Class), 
+                                 ggplot2::aes(Class, Count, group = 1)) + 
+              ggplot2::geom_point(colour = colour) + 
+              ggplot2::geom_line(colour = colour) + 
+              ggplot2::theme_classic() + 
               ggplot2::facet_wrap(~key, scales = "free_y", 
                                   ncol = facet.arrange)
   
           }
           else if (!is.null(samples)) {
-            counts.df <- counts.df %>% select(!all_of(samples))
-            p <- ggplot2::ggplot(tidyr::gather(counts.df, 
-                                               key, Count, -Class), ggplot2::aes(Class, 
-                                                                                 Count, group = 1)) + ggplot2::geom_point() + 
-              ggplot2::geom_line() + ggplot2::theme_classic() + 
-              ggplot2::facet_wrap(~key, scales = "free_y", 
-                                  ncol = facet.arrange)
+            counts_class <- counts.df %>% select(Class)
+            counts.df <- counts.df %>% select(all_of(samples))
+            counts.df <- cbind(counts_class,counts.df )
+            p <- ggplot2::ggplot(tidyr::gather(counts.df,  key, Count, -Class), 
+                                 ggplot2::aes(Class, Count, group = 1)) + 
+              ggplot2::geom_point(colour = colour) + 
+              ggplot2::geom_line(colour = colour) + 
+              ggplot2::theme_classic() + 
+              ggplot2::facet_wrap(~key, scales = "free_y", ncol = facet.arrange)
           }
           out <- list(plot = p, data = counts.df)
           return(out)
         }
         else if (facet == FALSE) {
-          plist2 = sapply(names(counts.df)[-grep("Class", 
-                                                 names(counts.df))], function(col) {
-                                                   ggplot2::ggplot(counts.df, ggplot2::aes_string(x = "Class", 
-                                                                                                  y = col, group = 1)) + ggplot2::geom_point(colour = colour) + 
-                                                     ggplot2::theme_classic() + ggplot2::geom_line(colour = colour) + 
-                                                     ggplot2::labs(title = col, x = "RNA Class", 
-                                                                   y = "Count")
-                                                 }, simplify = FALSE)
+          if(!is.null(samples)){
+            counts_class <- counts.df %>% select(Class)
+            counts.df <- counts.df %>% select(all_of(samples))
+            counts.df <- cbind(counts_class,counts.df )
+          }
+          plist2 = sapply(names(counts.df)[
+            -grep("Class", names(counts.df))], function(col) {
+              ggplot2::ggplot(counts.df, 
+                              ggplot2::aes_string(x ="Class",y =col,group = 1))+ 
+                ggplot2::geom_point(colour = colour) + 
+                ggplot2::theme_classic() + 
+                ggplot2::geom_line(colour = colour) + 
+                ggplot2::labs(title = col, x = "RNA Class", y = "Count")}, 
+            simplify = FALSE)
           sn <- names(plist2)
           p <- plist2
           out <- list(plot = p, data = counts.df)